Services/api/requirements.txt
uvicorn==0.18.3
pydantic==1.10.2
fastapi==0.85.0
SQLAlchemy==1.4.41
PyMySQL==1.0.2
passlib==1.7.4
bcrypt==4.0.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.1